requirements.txt
numpy==1.16.2
Flask==0.11.1
requests==2.22.0
Pillow==5.4.1
